# Implement graph construction as an llvm::ModulePass