--- a/deeplearning/ml4pl/graphs/llvm2graph/graph_builder_pass.cc
+++ b/deeplearning/ml4pl/graphs/llvm2graph/graph_builder_pass.cc
@@ -1,0 +1,29 @@
+// Copyright 2019 the ProGraML authors.
+//
+// Contact Chris Cummins <chrisc.101@gmail.com>.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#include "deeplearning/ml4pl/graphs/llvm2graph/graph_builder_pass.h"
+
+#include "labm8/cpp/statusor.h"
+
+namespace ml4pl {
+
+char GraphBuilderPass::ID = 0;
+
+bool GraphBuilderPass::runOnModule(llvm::Module& module) {
+  graph_ = graphBuilder_.Build(module);
+  return /*modified=*/false;
+}
+
+}  // namespace ml4pl