This implements the LLVM IR -> ProGraML graph protos as an llvm::ModulePass, rather than a big pile of Python string manipulation built on top of opt's dotfile output. The advantages of this are:

1. Faster - on the order of 1.8-900x (see benchmarks below).
2. More robust - no more fuzzy string matching and bugs from dotfile formatting.
3. Access to much richer information while constructing the graph - e.g. full type info, LLVM analysis passes, etc.

That last point provides some interesting future research directions. E.g. the graph representation currently encodes all of the operand type info for statements inside the statement embedding. A nicer way to do this would be to have the data node have an embedding for type information, then statement embeddings need only consider the opcode / modifiers, not the operands.

For now, there are a few packages which still depend on the old llvm2graph implementation (e.g. the polyhedra graph annotator), so I've kept the old implement and moved it into a legacy pacakge:

```sh
//deeplearning/ml4pl/graphs/llvm2graph  # the new LLVM-based implementation
//deeplearning/ml4pl/graphs/llvm2graph/legacy  # the old python/opt implementation
```

Once the last dependency on the legacy code has been refactored, I'll remove this pacakge.

## Usage

```sh
$ bazel run //deeplearning/ml4pl/graphs/llvm2graph -- /path/to/ir.ll
```

The program graph proto is printed to stdout. Alternatively, use `--stdout_format=dot` to print a graphviz graph:

```sh
$ bazel run //deeplearning/ml4pl/graphs/llvm2graph -- \
    $PWD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/53.ll \
    --stdout_format=dot > /tmp/graph.dot
$ dot -Tpng /tmp/graph.dot -o /tmp/graph.png
```

An example graph from the POJ-104 dataset:

![dot](https://user-images.githubusercontent.com/1636663/73076604-56d52280-3eb6-11ea-9390-d7aaddc111a4.png)

## Benchmarks

```sh
# old implementation:
$ time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/legacy/llvm2graph \
    < $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/53.ll
0.80s user 0.41s system 176% cpu 0.688 total

# new implementation:
$ time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/llvm2graph_ld_preload \
    $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/53.ll
0.04s user 0.01s system 76% cpu 0.068 total
```
**10x** speedup

```sh
# old implementation:
time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/legacy/llvm2graph \
    < $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/105975.ll
2.58s user 0.45s system 118% cpu 2.548 total

# new implementation:
time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/llvm2graph_ld_preload \
    $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/105975.ll
0.24s user 0.04s system 88% cpu 0.313 total
```
**8.1x** speedup

```sh
# old implementation:
time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/legacy/llvm2graph \
    < $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/115532.ll
11.84s user 0.52s system 102% cpu 12.075 total

# new implementation:
time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/llvm2graph_ld_preload \
    $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/115532.ll
6.32s user 0.21s system 94% cpu 6.866 total
```
**1.8x** speedup


```sh
# old implementation:
time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/legacy/llvm2graph \
    < $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/4180.ll
TimeoutError: Graph construction did not complete within 120 seconds

# new implementation:
$ time bazel-bin/deeplearning/ml4pl/graphs/llvm2graph/llvm2graph_ld_preload \
    $PHD/deeplearning/ml4pl/testing/data/bytecode_regression_tests/4180.ll
0.06s user 0.02s system 60% cpu 0.133 total
```
**> 902x** speedup

## References

github.com/ChrisCummins/ProGraML/issues/46