@ChrisCummins 2020-01-22
Todo: add intermediate data nodes between statements