@ChrisCummins 2020-01-24
Hey @Zacharias030 I wouldn't ask you to look through all 2.5k LOC diff but if you get a chance to have a quick skim over the main changes that would be great :) 

1. The old llvm2graph has been moved into a `legacy` pacakge, but kept unchanged.
1. New graph generator fixes a few issues with the old one (fuzzy resolving of callsites, requirements for functions to have an exit statement, etc), and can be run using `//deeplearning/ml4pl/graphs/llvm2graph`. There's a python wrapper which provides the same functionality as the old graph builder.
1. Improvements to the graphviz visualizer (bounding boxes around functions)